https://github.com/smallnest/gen/issues/17 was fixed in the following https://github.com/jimsmart/schema/issues/6
This is pulling the latest
	github.com/jimsmart/schema v0.0.0-20191104172732-39e4bec5335f
which has the mssql.Driver mapped in.

Further testing showed that there seems to be a difference between how mssql accepts a table from in the FROM statement.
i.e. FROM Vehicle and NOT FROM `Vehicle`

https://github.com/jimsmart/schema/issues/10

To be backwards compatible with the jimsmart/schema library I am passing in the right table format based on the sqlType